refactor: move default factory to internal
